# Conflicts:
#	src/main/java/com/space/munova/product/application/ProductDetailService.java

+++ src/main/java/com/space/munova/payment/dto/CancelDto.java
@@ -12,7 +12,7 @@ public record CancelDto(
         Integer transferDiscountAmount,
         Integer easyPayDiscountAmount,
         String receiptKey,
-        Integer cancelAmount,
+        Long cancelAmount,
         Integer taxFreeAmount,
         Integer refundableAmount,
         String cancelStatus,
